remove named_apply to flowvision.models.helpers
