+++ flowvision/models/mlp_mixer.py
@@ -5,12 +5,12 @@ import oneflow as flow
 import oneflow.nn as nn
 import oneflow.nn.init as init
 
-from .utils import load_state_dict_from_url, named_apply
-from .registry import ModelCreator
 from flowvision.layers.regularization import DropPath
 from flowvision.layers.blocks import PatchEmbed
 from flowvision.layers.weight_init import lecun_normal_
-
+from .utils import load_state_dict_from_url
+from .helpers import named_apply
+from .registry import ModelCreator
 
 model_urls = {
     "mlp_mixer_s16_224": None,
