+++ flowvision/models/utils.py
@@ -11,10 +11,8 @@ import warnings
 from urllib.parse import urlparse
 from urllib.request import Request, urlopen
 from tqdm import tqdm
-from typing import Any, Callable, Optional, Tuple
 
 import oneflow as flow
-import oneflow.nn as nn
 
 HASH_REGEX = re.compile(r"([a-f0-9]*)_")
 
@@ -184,37 +182,3 @@ def download_url_to_file(url, dst, hash_prefix=None, progress=True):
         f.close()
         if os.path.exists(f.name):
             os.remove(f.name)
-
-
-def named_apply(
-    fn: Callable, module: nn.Module, name="", depth_first=True, include_root=False
-) -> nn.Module:
-    if not depth_first and include_root:
-        fn(module=module, name=name)
-    for child_name, child_module in module.named_children():
-        child_name = ".".join((name, child_name)) if name else child_name
-        named_apply(
-            fn=fn,
-            module=child_module,
-            name=child_name,
-            depth_first=depth_first,
-            include_root=True,
-        )
-    if depth_first and include_root:
-        fn(module=module, name=name)
-    return module
-
-
-def named_modules(module: nn.Module, name="", depth_first=True, include_root=False):
-    if not depth_first and include_root:
-        yield name, module
-    for child_name, child_module in module.named_children():
-        child_name = ".".join((name, child_name)) if name else child_name
-        yield from named_modules(
-            module=child_module,
-            name=child_name,
-            depth_first=depth_first,
-            include_root=True,
-        )
-    if depth_first and include_root:
-        yield name, module
